// frontend/features/api/model/reportDebtAging200DataItem.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { ReportDebtAging200DataItemBuckets } from './reportDebtAging200DataItemBuckets';

export type ReportDebtAging200DataItem = {
  customer_id: string;
  customer_name: string;
  total_debt: number;
  buckets: ReportDebtAging200DataItemBuckets;
};